+++ R/recipes-step_ts_clean.R
@@ -138,12 +138,12 @@ prep.step_ts_clean <- function(x, training, info = NULL, ...) {
     recipes::check_type(training[, col_names])
 
     # Lambda Calculation
-    if (x$lambda[1] == "auto") {
+    if (is.null(x$lambda[1])) {
+        lambda_values <- rep(NA, length(col_names))
+        names(lambda_values) <- col_names
+    } else if (x$lambda[1] == "auto") {
         lambda_values <- training[, col_names] %>%
             purrr::map(auto_lambda)
-    } else if (is.null(x$lamda[1])) {
-        lambda_values <- rep(NULL, length(col_names))
-        names(lambda_values) <- col_names
     } else {
         lambda_values <- rep(x$lambda[1], length(col_names))
         names(lambda_values) <- col_names
@@ -172,7 +172,7 @@ bake.step_ts_clean <- function(object, new_data, ...) {
 
         # Handle "non-numeric" naming issue
         val_i <- object$lambdas_trained[i]
-        if (!is.null(val_i)) {
+        if (!is.na(val_i)) {
             val_i <- as.numeric(val_i)
         }